Close AG1-FIX-2 peer listener follow-ups
