--- crates/atm-daemon/src/peer_transport.rs
@@ -1,2371 +1,0 @@
-use std::io::{self, Read, Write};
-use std::net::{SocketAddr, TcpListener, TcpStream};
-use std::panic::{AssertUnwindSafe, catch_unwind};
-#[cfg(test)]
-use std::path::PathBuf;
-use std::sync::OnceLock;
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::mpsc::{self, Receiver, SyncSender};
-use std::sync::{Arc, Mutex};
-use std::thread;
-use std::time::{Duration, Instant};
-
-use atm_core::AtmConfig;
-use atm_core::boundary::{
-    self, AtmProtocol, ClientTransport, MessageKey, RemoteReplayStateRecord, RemoteReplayStore,
-    RequestDispatcher,
-};
-use atm_core::error::{AtmError, AtmErrorCode};
-use atm_core::protocol::{
-    JsonAtmProtocolCodec, RequestEnvelope, ResponseEnvelope, TeamMemberHeartbeatRequest,
-};
-use atm_core::types::{AgentName, IsoTimestamp, TeamName};
-
-use crate::active_connection_registry::{ActiveConnectionRegistry, TrackedDispatchHandle};
-use crate::{DaemonSubsystem, SubsystemObservability};
-
-// Architecture authority: docs/architecture.md §21.6.4 daemon operational
-// defaults and remote peer transport rules.
-// These deadlines are intentionally fixed operational constants. Phase Y keeps
-// connect and I/O timeouts non-configurable so remote peer delivery behavior
-// stays bounded and auditable across every host.
-const PEER_CONNECT_DEADLINE: Duration = Duration::from_secs(5);
-const PEER_IO_DEADLINE: Duration = Duration::from_secs(5);
-const DEFAULT_REMOTE_RETRY_BUDGET: Duration = Duration::from_secs(30);
-const MIN_REMOTE_RETRY_BUDGET: Duration = Duration::from_secs(1);
-const MAX_REMOTE_RETRY_BUDGET: Duration = Duration::from_secs(7 * 24 * 60 * 60);
-const INITIAL_RETRY_BACKOFF: Duration = Duration::from_millis(250);
-const MAX_RETRY_BACKOFF: Duration = Duration::from_secs(5);
-const MAX_REMOTE_REPLAY_RESUME_RECORDS: usize = 10_000;
-const PEER_LISTENER_ACCEPT_POLL_INTERVAL: Duration = Duration::from_millis(50);
-const PEER_REQUEST_DEADLINE: Duration = Duration::from_secs(3);
-const PEER_CONNECTION_IO_SLICE: Duration = Duration::from_millis(200);
-const PEER_LISTENER_SHUTDOWN_DEADLINE: Duration = Duration::from_secs(3);
-const PEER_ACCEPT_ERROR_RETRY_BACKOFF: Duration = Duration::from_millis(200);
-const MAX_CONCURRENT_PEER_CONNECTIONS: usize = 64;
-const MAX_TRACKED_PEER_DISPATCH_HANDLES: usize = 256;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) struct PeerTransportConfig {
-    pub(crate) remote_retry_budget: Duration,
-    pub(crate) peer_listen_addr: Option<SocketAddr>,
-}
-
-impl Default for PeerTransportConfig {
-    fn default() -> Self {
-        Self {
-            remote_retry_budget: DEFAULT_REMOTE_RETRY_BUDGET,
-            peer_listen_addr: None,
-        }
-    }
-}
-
-impl PeerTransportConfig {
-    pub(crate) fn from_config(config: Option<&AtmConfig>) -> Result<Self, AtmError> {
-        if config.is_none() {
-            tracing::warn!(
-                subsystem = "peer_transport",
-                action = "config_default",
-                outcome = "default",
-                "no AtmConfig provided to PeerTransportConfig::from_config; using default remote_retry_budget"
-            );
-        }
-        let remote_retry_budget = config
-            .map(|config| config.daemon.remote_retry_budget)
-            .unwrap_or(DEFAULT_REMOTE_RETRY_BUDGET);
-        if remote_retry_budget < MIN_REMOTE_RETRY_BUDGET {
-            return Err(AtmError::validation(format!(
-                "daemon.remote_retry_budget must be at least {} second(s)",
-                MIN_REMOTE_RETRY_BUDGET.as_secs()
-            ))
-            .with_recovery(
-                "Raise daemon.remote_retry_budget to at least one second before starting atm-daemon.",
-            ));
-        }
-        if remote_retry_budget > MAX_REMOTE_RETRY_BUDGET {
-            return Err(AtmError::validation(format!(
-                "daemon.remote_retry_budget must not exceed {} second(s)",
-                MAX_REMOTE_RETRY_BUDGET.as_secs()
-            ))
-            .with_recovery(
-                "Lower daemon.remote_retry_budget to one week or less before starting atm-daemon.",
-            ));
-        }
-        Ok(Self {
-            remote_retry_budget,
-            peer_listen_addr: config.and_then(|config| config.daemon.peer_listen_addr),
-        })
-    }
-}
-
-fn remote_replay_store_not_configured_error() -> AtmError {
-    AtmError::daemon_unavailable("remote replay store is not configured").with_recovery(
-        "Restore the host-scoped ATM durable replay store before retrying remote delivery so atm-daemon can resume unknown peer handoffs safely.",
-    )
-}
-
-fn remote_peer_endpoint_not_configured_error() -> AtmError {
-    AtmError::daemon_unavailable("remote peer endpoint is not configured").with_recovery(
-        "Set ATM_DAEMON_PEER_ADDR or configure the daemon peer transport before retrying remote delivery or replay persistence.",
-    )
-}
-
-fn remote_retry_budget_expiry_error(
-    source: impl std::error::Error + Send + Sync + 'static,
-) -> AtmError {
-    AtmError::daemon_unavailable("failed to convert remote retry budget into a replay expiry")
-        .with_recovery(
-            "Fix the daemon remote retry budget configuration and restart atm-daemon before retrying remote delivery so replay expiry can be computed deterministically.",
-        )
-        .with_source(source)
-}
-
-fn remote_replay_persistence_failed_error(source: AtmError) -> AtmError {
-    AtmError::remote_delivery_outcome_unknown(
-        "remote peer delivery outcome is unknown and replay persistence failed",
-    )
-    .with_source(source)
-}
-
-fn daemon_peer_endpoint_from_env() -> Option<SocketAddr> {
-    match std::env::var("ATM_DAEMON_PEER_ADDR") {
-        Ok(raw) => parse_peer_endpoint(&raw),
-        Err(std::env::VarError::NotPresent) => None,
-        Err(std::env::VarError::NotUnicode(_)) => {
-            tracing::warn!(
-                subsystem = "peer_transport",
-                action = "env_parse",
-                outcome = "ignored",
-                "ignoring non-unicode ATM_DAEMON_PEER_ADDR value"
-            );
-            None
-        }
-    }
-}
-
-#[derive(Debug)]
-struct PeerServerHandle {
-    terminate: Arc<AtomicBool>,
-    join_handle: std::thread::JoinHandle<Result<(), AtmError>>,
-    #[cfg_attr(not(test), allow(dead_code))]
-    bound_addr: SocketAddr,
-}
-
-#[derive(Debug)]
-struct PeerServerTransport {
-    listen_addr: Mutex<Option<SocketAddr>>,
-    observability: SubsystemObservability,
-    state: Mutex<Option<PeerServerHandle>>,
-}
-
-impl PeerServerTransport {
-    fn new(listen_addr: Option<SocketAddr>, observability: SubsystemObservability) -> Self {
-        Self {
-            listen_addr: Mutex::new(listen_addr),
-            observability,
-            state: Mutex::new(None),
-        }
-    }
-
-    fn start(&self, dispatcher: Arc<dyn RequestDispatcher + Send + Sync>) -> Result<(), AtmError> {
-        let listen_addr = *self.listen_addr.lock().map_err(|_| {
-            AtmError::daemon_unavailable("peer listener config lock poisoned")
-                .with_recovery("Restart atm-daemon before retrying cross-host peer startup.")
-        })?;
-        let Some(listen_addr) = listen_addr else {
-            return Ok(());
-        };
-        let mut state = self.state.lock().map_err(|_| {
-            AtmError::daemon_unavailable("peer listener state lock poisoned")
-                .with_recovery("Restart atm-daemon before retrying cross-host peer startup.")
-        })?;
-        if state.is_some() {
-            return Ok(());
-        }
-        let listener = TcpListener::bind(listen_addr).map_err(|source| {
-            AtmError::daemon_unavailable(format!(
-                "failed to bind daemon peer listener at {listen_addr}"
-            ))
-            .with_recovery(
-                "Choose an available literal IP:port for daemon.peer_listen_addr and restart atm-daemon.",
-            )
-            .with_source(source)
-        })?;
-        listener.set_nonblocking(true).map_err(|source| {
-            AtmError::daemon_unavailable("failed to configure daemon peer listener as nonblocking")
-                .with_recovery(
-                    "Restart atm-daemon after confirming the host allows nonblocking TCP listeners for cross-host peer transport.",
-                )
-                .with_source(source)
-        })?;
-        let bound_addr = listener.local_addr().map_err(|source| {
-            AtmError::daemon_unavailable("failed to resolve daemon peer listener bound address")
-                .with_recovery(
-                    "Restart atm-daemon after confirming the configured daemon.peer_listen_addr is valid on this host.",
-                )
-                .with_source(source)
-        })?;
-        let terminate = Arc::new(AtomicBool::new(false));
-        let terminate_for_thread = Arc::clone(&terminate);
-        let observability = self.observability.clone();
-        let join_handle = thread::Builder::new()
-            .name("peer-transport-listener".to_string())
-            .spawn(move || {
-                serve_peer_listener(listener, dispatcher, terminate_for_thread, observability)
-            })
-            .map_err(|source| {
-                AtmError::daemon_unavailable("failed to spawn daemon peer listener thread")
-                    .with_recovery(
-                        "Restart atm-daemon after confirming the host can create the cross-host peer listener worker.",
-                    )
-                    .with_source(source)
-            })?;
-        self.observability.emit_or_warn(
-            "peer_listener_start",
-            "ok",
-            format!("daemon peer listener bound at {bound_addr}"),
-        );
-        *state = Some(PeerServerHandle {
-            terminate,
-            join_handle,
-            bound_addr,
-        });
-        Ok(())
-    }
-
-    fn shutdown(&self) -> Result<(), AtmError> {
-        let handle = self
-            .state
-            .lock()
-            .map_err(|_| {
-                AtmError::daemon_unavailable("peer listener state lock poisoned")
-                    .with_recovery("Restart atm-daemon before retrying cross-host peer shutdown.")
-            })?
-            .take();
-        let Some(handle) = handle else {
-            return Ok(());
-        };
-        handle.terminate.store(true, Ordering::SeqCst);
-        let (result_tx, result_rx) = mpsc::sync_channel(1);
-        thread::Builder::new()
-            .name("peer-transport-join-helper".to_string())
-            .spawn(move || {
-                let result = match handle.join_handle.join() {
-                    Ok(Ok(())) => Ok(()),
-                    Ok(Err(error)) => Err(error),
-                    Err(_) => Err(AtmError::daemon_unavailable(
-                        "daemon peer listener thread panicked unexpectedly",
-                    )
-                    .with_recovery(
-                        "Restart atm-daemon before retrying cross-host peer transport.",
-                    )),
-                };
-                let _ = result_tx.send(result);
-            })
-            .map_err(|source| {
-                AtmError::daemon_unavailable("failed to spawn peer listener join helper")
-                    .with_recovery(
-                        "Restart atm-daemon before retrying cross-host peer listener shutdown.",
-                    )
-                    .with_source(source)
-            })?;
-        match result_rx.recv_timeout(PEER_LISTENER_SHUTDOWN_DEADLINE) {
-            Ok(result) => result,
-            Err(mpsc::RecvTimeoutError::Timeout) => Err(AtmError::daemon_unavailable(
-                "daemon peer listener shutdown exceeded the bounded join deadline",
-            )
-            .with_recovery(
-                "Restart atm-daemon after confirming no inbound peer connection is stalled beyond the bounded deadline.",
-            )),
-            Err(mpsc::RecvTimeoutError::Disconnected) => Err(AtmError::daemon_unavailable(
-                "daemon peer listener join helper disconnected unexpectedly",
-            )
-            .with_recovery(
-                "Restart atm-daemon before retrying cross-host peer listener shutdown.",
-            )),
-        }
-    }
-
-    fn reload(
-        &self,
-        listen_addr: Option<SocketAddr>,
-        dispatcher: Arc<dyn RequestDispatcher + Send + Sync>,
-    ) -> Result<(), AtmError> {
-        let previous = {
-            let mut configured = self.listen_addr.lock().map_err(|_| {
-                AtmError::daemon_unavailable("peer listener config lock poisoned").with_recovery(
-                    "Restart atm-daemon before retrying cross-host peer listener reload.",
-                )
-            })?;
-            let previous = *configured;
-            if previous == listen_addr {
-                return Ok(());
-            }
-            *configured = listen_addr;
-            previous
-        };
-
-        if previous.is_some() {
-            self.shutdown()?;
-        }
-        if listen_addr.is_some() {
-            self.start(dispatcher).inspect_err(|error| {
-                tracing::warn!(
-                    subsystem = "peer_transport",
-                    action = "reload_listener",
-                    outcome = "degraded",
-                    %error,
-                    "daemon peer listener reload failed to rebind the configured address"
-                );
-                self.observability.emit_or_warn(
-                    "peer_listener_reload",
-                    "degraded",
-                    "daemon peer listener reload failed to rebind the configured address",
-                );
-            })?;
-        }
-        Ok(())
-    }
-
-    #[cfg(test)]
-    fn bound_addr_for_test(&self) -> Option<SocketAddr> {
-        self.state
-            .lock()
-            .ok()
-            .and_then(|state| state.as_ref().map(|handle| handle.bound_addr))
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum AttemptFailureKind {
-    Retryable,
-    NonRetryable,
-    OutcomeUnknown,
-}
-
-#[derive(Debug)]
-struct AttemptFailure {
-    kind: AttemptFailureKind,
-    error: AtmError,
-}
-
-struct DeliveryRetryState<'a> {
-    deadline: Instant,
-    terminate: &'a Arc<AtomicBool>,
-    backoff: &'a mut Duration,
-    next_attempt: &'a mut u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) struct ReplayResumeSummary {
-    pub(crate) delivered: usize,
-    pub(crate) retained: usize,
-    pub(crate) purged_expired: usize,
-}
-
-#[derive(Clone)]
-struct PeerClientTransport {
-    endpoint: Option<SocketAddr>,
-    config: PeerTransportConfig,
-    replay_store: Option<Arc<dyn RemoteReplayStore>>,
-    codec: JsonAtmProtocolCodec,
-    observability: SubsystemObservability,
-}
-
-impl std::fmt::Debug for PeerClientTransport {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("PeerClientTransport")
-            .field("endpoint", &self.endpoint)
-            .field("config", &self.config)
-            .field(
-                "replay_store",
-                &self.replay_store.as_ref().map(|_| "dyn RemoteReplayStore"),
-            )
-            .field("codec", &"JsonAtmProtocolCodec")
-            .field("observability", &self.observability)
-            .finish()
-    }
-}
-
-impl PeerClientTransport {
-    fn new_with_observability(
-        replay_store: Option<Arc<dyn RemoteReplayStore>>,
-        config: PeerTransportConfig,
-        observability: SubsystemObservability,
-    ) -> Self {
-        let endpoint = daemon_peer_endpoint_from_env();
-        Self {
-            endpoint,
-            config,
-            replay_store,
-            codec: JsonAtmProtocolCodec,
-            observability,
-        }
-    }
-
-    #[cfg(test)]
-    fn new_for_test(
-        endpoint: SocketAddr,
-        config: PeerTransportConfig,
-        replay_db_path: PathBuf,
-    ) -> Self {
-        let replay_store =
-            atm_runtime::sqlite_remote_replay_store_for_test(replay_db_path).expect("replay store");
-        Self {
-            endpoint: Some(endpoint),
-            config,
-            replay_store: Some(replay_store),
-            codec: JsonAtmProtocolCodec,
-            observability: SubsystemObservability::disabled(DaemonSubsystem::PeerTransport),
-        }
-    }
-
-    fn resume_pending_replay(&self) -> Result<ReplayResumeSummary, AtmError> {
-        let Some(replay_store) = &self.replay_store else {
-            return Ok(ReplayResumeSummary {
-                delivered: 0,
-                retained: 0,
-                purged_expired: 0,
-            });
-        };
-
-        let now = IsoTimestamp::now();
-        let purged_expired = replay_store.purge_expired(now)?;
-        let records = replay_store.load_all()?;
-        if records.len() > MAX_REMOTE_REPLAY_RESUME_RECORDS {
-            return Err(AtmError::daemon_unavailable(format!(
-                "daemon remote replay resume exceeded the bounded record cap ({MAX_REMOTE_REPLAY_RESUME_RECORDS})"
-            ))
-            .with_recovery(
-                "Drain or delete retained remote replay rows until the bounded startup replay cap is back under control, then restart atm-daemon.",
-            ));
-        }
-        let mut delivered = 0usize;
-        let mut retained = 0usize;
-        for mut record in records {
-            match self.send_to_endpoint(record.peer_addr, record.request.clone()) {
-                Ok(_) => {
-                    replay_store.delete(&record.team, &record.agent, &record.message_key)?;
-                    tracing::info!(
-                        message_key = %record.message_key,
-                        peer_addr = %record.peer_addr,
-                        replay_attempt_count = record.attempt_count,
-                        "daemon remote replay delivered successfully"
-                    );
-                    self.observability.emit_or_warn(
-                        "resume_pending_replay",
-                        "ok",
-                        "daemon remote replay delivered a retained record",
-                    );
-                    delivered += 1;
-                }
-                Err(error) => {
-                    record.attempt_count = record.attempt_count.saturating_add(1);
-                    record.last_attempt_at = Some(IsoTimestamp::now());
-                    record.last_error = Some(error.code);
-                    tracing::warn!(
-                        subsystem = "peer_transport",
-                        action = "resume_replay",
-                        outcome = "skipped",
-                        message_key = %record.message_key,
-                        peer_addr = %record.peer_addr,
-                        replay_attempt_count = record.attempt_count,
-                        error_code = %error.code,
-                        error_message = %error.message,
-                        "daemon remote replay delivery attempt failed; retaining record"
-                    );
-                    self.observability.emit_or_warn(
-                        "resume_pending_replay",
-                        "degraded",
-                        "daemon remote replay delivery failed and retained the record for retry",
-                    );
-                    replay_store.enqueue(record)?;
-                    retained += 1;
-                }
-            }
-        }
-
-        Ok(ReplayResumeSummary {
-            delivered,
-            retained,
-            purged_expired,
-        })
-    }
-
-    fn persist_replay_request(
-        &self,
-        team: TeamName,
-        agent: AgentName,
-        message_key: MessageKey,
-        request: RequestEnvelope,
-    ) -> Result<(), AtmError> {
-        let Some(replay_store) = &self.replay_store else {
-            return Err(remote_replay_store_not_configured_error());
-        };
-        let endpoint = self
-            .endpoint
-            .ok_or_else(remote_peer_endpoint_not_configured_error)?;
-        let recorded_at = IsoTimestamp::now();
-        // The configured retry budget intentionally bounds both the live retry
-        // window and the retained replay TTL so durable replay cannot outlive
-        // the operator-visible retry promise for one remote delivery attempt.
-        let expires_at = IsoTimestamp::from_datetime(
-            recorded_at.into_inner()
-                + chrono::Duration::from_std(self.config.remote_retry_budget)
-                    .map_err(remote_retry_budget_expiry_error)?,
-        );
-        replay_store.enqueue(RemoteReplayStateRecord {
-            team,
-            agent,
-            message_key,
-            peer_addr: endpoint,
-            request,
-            recorded_at,
-            expires_at,
-            attempt_count: 0,
-            last_attempt_at: None,
-            last_error: None,
-        })
-    }
-
-    fn persist_outcome_unknown_request(&self, request: &RequestEnvelope) -> Result<(), AtmError> {
-        let Some((team, agent, message_key)) = replay_metadata_for_request(request) else {
-            tracing::warn!(
-                subsystem = "peer_transport",
-                action = "persist",
-                outcome = "unknown",
-                request = ?request,
-                "remote delivery outcome is unknown but this request family does not support durable replay persistence",
-            );
-            // The original RemoteDeliveryOutcomeUnknown already carries the shared operator
-            // guidance for retry-vs-resume, so unsupported request families intentionally avoid
-            // fabricating a second replay-specific wrapper here.
-            return Ok(());
-        };
-        self.persist_replay_request(team, agent, message_key, request.clone())
-    }
-
-    fn send_to_endpoint(
-        &self,
-        endpoint: SocketAddr,
-        request: RequestEnvelope,
-    ) -> Result<ResponseEnvelope, AtmError> {
-        let frame = self
-            .codec
-            .request_to_frame(atm_core::protocol::next_request_id(), request)?;
-        let deadline = Instant::now() + self.config.remote_retry_budget;
-        let terminate = daemon_terminate_flag()?;
-        let mut backoff = INITIAL_RETRY_BACKOFF;
-        let mut attempt = 0u32;
-
-        loop {
-            self.ensure_retry_not_terminated(
-                &terminate,
-                "daemon shutdown interrupted remote peer delivery before the next network attempt",
-            )?;
-            let current_attempt = attempt;
-            let mut retry_state = DeliveryRetryState {
-                deadline,
-                terminate: &terminate,
-                backoff: &mut backoff,
-                next_attempt: &mut attempt,
-            };
-            match self.send_once(endpoint, &frame) {
-                Ok(response) => {
-                    return Ok(self.record_send_success(endpoint, current_attempt, response));
-                }
-                Err(failure) => {
-                    match self.handle_send_failure(
-                        endpoint,
-                        current_attempt,
-                        &mut retry_state,
-                        *failure,
-                    ) {
-                        DeliveryLoopDecision::Retry => {}
-                        DeliveryLoopDecision::Return(error) => return Err(error),
-                    }
-                }
-            }
-        }
-    }
-
-    fn send_once(
-        &self,
-        endpoint: SocketAddr,
-        request_frame: &atm_core::protocol::FramePayload,
-    ) -> Result<ResponseEnvelope, Box<AttemptFailure>> {
-        let mut stream = self.connect_peer_stream(endpoint)?;
-        self.apply_peer_io_deadlines(&stream)?;
-        self.publish_request_frame(&mut stream, request_frame)?;
-        match self.decode_response_frame(
-            request_frame.request_id,
-            self.read_response_frame(&mut stream)?,
-        )? {
-            ResponseEnvelope::Error(error) => Err(Box::new(AttemptFailure {
-                kind: AttemptFailureKind::NonRetryable,
-                error: error.into_atm_error(),
-            })),
-            response => Ok(response),
-        }
-    }
-
-    fn connect_peer_stream(&self, endpoint: SocketAddr) -> Result<TcpStream, Box<AttemptFailure>> {
-        TcpStream::connect_timeout(&endpoint, PEER_CONNECT_DEADLINE).map_err(|source| {
-            Box::new(AttemptFailure {
-                kind: classify_io_error(&source),
-                error: AtmError::daemon_unavailable(format!(
-                    "failed to connect to remote daemon peer at {endpoint}"
-                ))
-                .with_recovery(
-                    "Confirm the remote daemon is reachable at the configured peer endpoint, then retry. If the remote daemon is intentionally offline, let durable replay resume the handoff after it recovers.",
-                )
-                .with_source(source),
-            })
-        })
-    }
-
-    fn apply_peer_io_deadlines(&self, stream: &TcpStream) -> Result<(), Box<AttemptFailure>> {
-        stream
-            .set_read_timeout(Some(PEER_IO_DEADLINE))
-            .map_err(peer_read_deadline_error)?;
-        stream
-            .set_write_timeout(Some(PEER_IO_DEADLINE))
-            .map_err(peer_write_deadline_error)?;
-        Ok(())
-    }
-
-    fn publish_request_frame(
-        &self,
-        stream: &mut TcpStream,
-        request_frame: &atm_core::protocol::FramePayload,
-    ) -> Result<(), Box<AttemptFailure>> {
-        atm_core::protocol::write_frame(
-            stream,
-            request_frame,
-            "failed to write remote peer request frame",
-        )
-        .map_err(|error| {
-            Box::new(AttemptFailure {
-                kind: AttemptFailureKind::Retryable,
-                error,
-            })
-        })?;
-        stream.flush().map_err(peer_flush_error)?;
-        Ok(())
-    }
-
-    fn read_response_frame(
-        &self,
-        stream: &mut TcpStream,
-    ) -> Result<atm_core::protocol::FramePayload, Box<AttemptFailure>> {
-        atm_core::protocol::read_frame(
-            stream,
-            "failed to read remote peer response frame",
-            "remote peer response frame exceeded the maximum supported size",
-        )
-        .map_err(|error| {
-            Box::new(AttemptFailure {
-                kind: AttemptFailureKind::OutcomeUnknown,
-                error,
-            })
-        })?
-        .ok_or_else(peer_closed_before_response_error)
-    }
-
-    fn decode_response_frame(
-        &self,
-        request_id: atm_core::protocol::RequestId,
-        response_frame: atm_core::protocol::FramePayload,
-    ) -> Result<ResponseEnvelope, Box<AttemptFailure>> {
-        let (response_id, response) = self
-            .codec
-            .response_from_frame(response_frame)
-            .map_err(peer_response_decode_error)?;
-        if response_id != request_id {
-            return Err(peer_response_id_mismatch_error(response_id, request_id));
-        }
-        Ok(response)
-    }
-
-    fn ensure_retry_not_terminated(
-        &self,
-        terminate: &Arc<AtomicBool>,
-        message: &'static str,
-    ) -> Result<(), AtmError> {
-        if terminate.load(Ordering::SeqCst) {
-            return Err(AtmError::daemon_unavailable(message).with_recovery(
-                "Retry the daemon operation after atm-daemon restarts and resumes pending remote replay work.",
-            ));
-        }
-        Ok(())
-    }
-
-    fn record_send_success(
-        &self,
-        endpoint: SocketAddr,
-        attempt: u32,
-        response: ResponseEnvelope,
-    ) -> ResponseEnvelope {
-        tracing::info!(
-            peer_addr = %endpoint,
-            attempt,
-            "daemon peer delivery succeeded"
-        );
-        self.observability
-            .emit_or_warn("send_to_endpoint", "ok", "daemon peer delivery succeeded");
-        response
-    }
-
-    fn handle_send_failure(
-        &self,
-        endpoint: SocketAddr,
-        attempt: u32,
-        retry_state: &mut DeliveryRetryState<'_>,
-        failure: AttemptFailure,
-    ) -> DeliveryLoopDecision {
-        match failure.kind {
-            AttemptFailureKind::Retryable => {
-                self.handle_retryable_failure(endpoint, attempt, retry_state, failure.error)
-            }
-            AttemptFailureKind::NonRetryable | AttemptFailureKind::OutcomeUnknown => {
-                self.handle_terminal_failure(endpoint, attempt, failure)
-            }
-        }
-    }
-
-    fn handle_retryable_failure(
-        &self,
-        endpoint: SocketAddr,
-        attempt: u32,
-        retry_state: &mut DeliveryRetryState<'_>,
-        error: AtmError,
-    ) -> DeliveryLoopDecision {
-        let now = Instant::now();
-        if now >= retry_state.deadline {
-            tracing::error!(
-                subsystem = "peer_transport",
-                action = "send_to_endpoint",
-                outcome = "retry_exhausted",
-                peer_addr = %endpoint,
-                attempt,
-                error_code = %error.code,
-                error_message = %error.message,
-                "daemon peer delivery exhausted retry budget"
-            );
-            self.observability.emit_or_warn(
-                "send_to_endpoint",
-                "failed",
-                "daemon peer delivery exhausted its retry budget",
-            );
-            return DeliveryLoopDecision::Return(error);
-        }
-        let remaining = retry_state.deadline.saturating_duration_since(now);
-        let sleep_for =
-            jittered_backoff(*retry_state.backoff, jitter_seed(endpoint, attempt)).min(remaining);
-        tracing::warn!(
-            subsystem = "peer_transport",
-            action = "retry",
-            outcome = "retrying",
-            peer_addr = %endpoint,
-            attempt,
-            sleep_ms = sleep_for.as_millis(),
-            error_code = %error.code,
-            error_message = %error.message,
-            "daemon peer delivery hit retryable failure"
-        );
-        self.observability.emit_or_warn(
-            "send_to_endpoint",
-            "degraded",
-            "daemon peer delivery hit a retryable failure",
-        );
-        if wait_for_retry_backoff(retry_state.terminate, sleep_for) {
-            return DeliveryLoopDecision::Return(
-                AtmError::daemon_unavailable(
-                    "daemon shutdown interrupted remote peer retry backoff",
-                )
-                .with_recovery(
-                    "Retry the daemon operation after atm-daemon restarts and resumes pending remote replay work.",
-                ),
-            );
-        }
-        *retry_state.backoff = retry_state.backoff.saturating_mul(2).min(MAX_RETRY_BACKOFF);
-        *retry_state.next_attempt = retry_state.next_attempt.saturating_add(1);
-        DeliveryLoopDecision::Retry
-    }
-
-    fn handle_terminal_failure(
-        &self,
-        endpoint: SocketAddr,
-        attempt: u32,
-        failure: AttemptFailure,
-    ) -> DeliveryLoopDecision {
-        let failure_kind = match failure.kind {
-            AttemptFailureKind::OutcomeUnknown => "outcome_unknown",
-            AttemptFailureKind::NonRetryable => "non_retryable",
-            AttemptFailureKind::Retryable => "retryable",
-        };
-        tracing::error!(
-            subsystem = "peer_transport",
-            action = "send_to_endpoint",
-            outcome = "terminal_failure",
-            peer_addr = %endpoint,
-            attempt,
-            failure_kind,
-            error_code = %failure.error.code,
-            error_message = %failure.error.message,
-            "daemon peer delivery failed"
-        );
-        self.observability.emit_or_warn(
-            "send_to_endpoint",
-            "failed",
-            "daemon peer delivery failed with a non-retryable or outcome-unknown error",
-        );
-        DeliveryLoopDecision::Return(failure.error)
-    }
-}
-
-enum DeliveryLoopDecision {
-    Retry,
-    Return(AtmError),
-}
-
-fn peer_read_deadline_error(source: io::Error) -> Box<AttemptFailure> {
-    Box::new(AttemptFailure {
-        kind: AttemptFailureKind::Retryable,
-        error: AtmError::daemon_unavailable("failed to apply remote peer read deadline")
-            .with_recovery(
-                "Restart atm-daemon and retry after the peer socket can accept bounded read deadlines again.",
-            )
-            .with_source(source),
-    })
-}
-
-fn peer_write_deadline_error(source: io::Error) -> Box<AttemptFailure> {
-    Box::new(AttemptFailure {
-        kind: AttemptFailureKind::Retryable,
-        error: AtmError::daemon_unavailable("failed to apply remote peer write deadline")
-            .with_recovery(
-                "Restart atm-daemon and retry after the peer socket can accept bounded write deadlines again.",
-            )
-            .with_source(source),
-    })
-}
-
-fn peer_flush_error(source: io::Error) -> Box<AttemptFailure> {
-    Box::new(AttemptFailure {
-        kind: AttemptFailureKind::OutcomeUnknown,
-        error: AtmError::remote_delivery_outcome_unknown(
-            "failed to flush the remote peer request frame before waiting for a response",
-        )
-        .with_source(source),
-    })
-}
-
-fn peer_closed_before_response_error() -> Box<AttemptFailure> {
-    Box::new(AttemptFailure {
-        kind: AttemptFailureKind::OutcomeUnknown,
-        error: AtmError::remote_delivery_outcome_unknown(
-            "remote peer closed the connection before returning a response frame",
-        )
-        .with_recovery(
-            "Check the destination daemon or mailbox before retrying. If local durable replay is enabled, let the daemon resume the pending handoff rather than guessing success.",
-        ),
-    })
-}
-
-fn peer_response_decode_error(error: AtmError) -> Box<AttemptFailure> {
-    Box::new(AttemptFailure {
-        kind: AttemptFailureKind::NonRetryable,
-        error: AtmError::daemon_unavailable("failed to decode remote peer response frame")
-            .with_recovery(
-                "Align the peer daemon builds so both sides speak the same ATM daemon protocol before retrying the remote delivery.",
-            )
-            .with_source(error),
-    })
-}
-
-fn peer_response_id_mismatch_error(
-    response_id: atm_core::protocol::RequestId,
-    request_id: atm_core::protocol::RequestId,
-) -> Box<AttemptFailure> {
-    Box::new(AttemptFailure {
-        kind: AttemptFailureKind::NonRetryable,
-        error: AtmError::daemon_unavailable(format!(
-            "remote peer response request_id {} did not match request_id {}",
-            response_id, request_id
-        ))
-        .with_recovery(
-            "Align the peer daemon builds so both sides use the same ATM daemon protocol contract before retrying.",
-        ),
-    })
-}
-
-fn wait_for_retry_backoff(terminate: &Arc<AtomicBool>, sleep_for: Duration) -> bool {
-    const RETRY_POLL_INTERVAL: Duration = Duration::from_millis(25);
-
-    let started = Instant::now();
-    loop {
-        if terminate.load(Ordering::SeqCst) {
-            return true;
-        }
-        let elapsed = started.elapsed();
-        if elapsed >= sleep_for {
-            return false;
-        }
-        let remaining = sleep_for.saturating_sub(elapsed);
-        thread::sleep(remaining.min(RETRY_POLL_INTERVAL));
-    }
-}
-
-fn daemon_terminate_flag() -> Result<Arc<AtomicBool>, AtmError> {
-    // Global terminate flag for signal handler: the hot path only needs a
-    // shared cached flag, so OnceLock avoids a per-call mutex while leaving
-    // LifecycleControlSourceAdapter::install() as the fallible owner of setup.
-    static DAEMON_TERMINATE_FLAG: OnceLock<Arc<AtomicBool>> = OnceLock::new();
-
-    if let Some(flag) = DAEMON_TERMINATE_FLAG.get() {
-        return Ok(Arc::clone(flag));
-    }
-
-    let flag = crate::lifecycle_control::LifecycleControlSourceAdapter::install()?.terminate_flag();
-    match DAEMON_TERMINATE_FLAG.set(Arc::clone(&flag)) {
-        Ok(()) => Ok(flag),
-        Err(existing) => Ok(existing),
-    }
-}
-
-impl boundary::sealed::Sealed for PeerClientTransport {}
-
-impl ClientTransport for PeerClientTransport {
-    fn send(&self, request: RequestEnvelope) -> Result<ResponseEnvelope, AtmError> {
-        let endpoint = self
-            .endpoint
-            .ok_or_else(remote_peer_endpoint_not_configured_error)?;
-        match self.send_to_endpoint(endpoint, request.clone()) {
-            Ok(response) => Ok(response),
-            Err(error) if error.code == AtmErrorCode::RemoteDeliveryOutcomeUnknown => {
-                self.persist_outcome_unknown_request(&request)
-                    .map_err(remote_replay_persistence_failed_error)?;
-                Err(error)
-            }
-            Err(error) => Err(error),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct PeerTransportRuntime {
-    client: PeerClientTransport,
-    server: Arc<PeerServerTransport>,
-}
-
-impl Default for PeerTransportRuntime {
-    fn default() -> Self {
-        Self::new(None)
-    }
-}
-
-impl PeerTransportRuntime {
-    pub(crate) fn new(replay_store: Option<Arc<dyn RemoteReplayStore>>) -> Self {
-        Self::new_with_observability(
-            replay_store,
-            PeerTransportConfig::default(),
-            SubsystemObservability::disabled(DaemonSubsystem::PeerTransport),
-        )
-    }
-
-    pub(crate) fn new_with_observability(
-        replay_store: Option<Arc<dyn RemoteReplayStore>>,
-        config: PeerTransportConfig,
-        observability: SubsystemObservability,
-    ) -> Self {
-        Self {
-            server: Arc::new(PeerServerTransport::new(
-                config.peer_listen_addr,
-                observability.clone(),
-            )),
-            client: PeerClientTransport::new_with_observability(
-                replay_store,
-                config,
-                observability,
-            ),
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn client_transport(&self) -> &dyn ClientTransport {
-        &self.client
-    }
-
-    pub(crate) fn resume_pending_replay(&self) -> Result<ReplayResumeSummary, AtmError> {
-        self.client.resume_pending_replay()
-    }
-
-    pub(crate) fn start(
-        &self,
-        dispatcher: Arc<dyn RequestDispatcher + Send + Sync>,
-    ) -> Result<(), AtmError> {
-        self.server.start(dispatcher)
-    }
-
-    pub(crate) fn shutdown(&self) -> Result<(), AtmError> {
-        self.server.shutdown()
-    }
-
-    pub(crate) fn reload_listener(
-        &self,
-        listen_addr: Option<SocketAddr>,
-        dispatcher: Arc<dyn RequestDispatcher + Send + Sync>,
-    ) -> Result<(), AtmError> {
-        self.server.reload(listen_addr, dispatcher)
-    }
-
-    #[cfg(test)]
-    pub(crate) fn new_for_test(
-        endpoint: SocketAddr,
-        config: PeerTransportConfig,
-        replay_db_path: PathBuf,
-    ) -> Self {
-        Self {
-            server: Arc::new(PeerServerTransport::new(
-                None,
-                SubsystemObservability::disabled(DaemonSubsystem::PeerTransport),
-            )),
-            client: PeerClientTransport::new_for_test(endpoint, config, replay_db_path),
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn new_server_for_test(
-        listen_addr: SocketAddr,
-        observability: SubsystemObservability,
-    ) -> Self {
-        let config = PeerTransportConfig {
-            remote_retry_budget: DEFAULT_REMOTE_RETRY_BUDGET,
-            peer_listen_addr: Some(listen_addr),
-        };
-        Self {
-            server: Arc::new(PeerServerTransport::new(
-                Some(listen_addr),
-                observability.clone(),
-            )),
-            client: PeerClientTransport::new_with_observability(None, config, observability),
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn bound_addr_for_test(&self) -> Option<SocketAddr> {
-        self.server.bound_addr_for_test()
-    }
-
-    #[cfg(test)]
-    pub(crate) fn persist_replay_request(
-        &self,
-        team: TeamName,
-        agent: AgentName,
-        message_key: MessageKey,
-        request: RequestEnvelope,
-    ) -> Result<(), AtmError> {
-        self.client
-            .persist_replay_request(team, agent, message_key, request)
-    }
-
-    #[cfg(test)]
-    pub(crate) fn load_pending_replay_records(
-        &self,
-    ) -> Result<Vec<RemoteReplayStateRecord>, AtmError> {
-        match &self.client.replay_store {
-            Some(replay_store) => replay_store.load_all(),
-            None => Ok(Vec::new()),
-        }
-    }
-}
-
-fn serve_peer_listener(
-    listener: TcpListener,
-    dispatcher: Arc<dyn RequestDispatcher + Send + Sync>,
-    terminate: Arc<AtomicBool>,
-    observability: SubsystemObservability,
-) -> Result<(), AtmError> {
-    let registry = Arc::new(ActiveConnectionRegistry::default());
-    thread::scope(|scope| {
-        while !terminate.load(Ordering::SeqCst) {
-            let reap_summary = registry.reap_finished_dispatches()?;
-            if reap_summary.recovered_panics > 0 {
-                observability.emit_or_warn(
-                    "peer_dispatch_worker",
-                    "panic_recovered",
-                    "peer dispatch worker panicked; listener recovered and continued",
-                );
-            }
-            match listener.accept() {
-                Ok((stream, peer_addr)) => {
-                    if registry.active_connections() >= MAX_CONCURRENT_PEER_CONNECTIONS {
-                        tracing::warn!(
-                            subsystem = "peer_transport",
-                            action = "accept_connection",
-                            outcome = "rejected",
-                            %peer_addr,
-                            max_concurrent_connections = MAX_CONCURRENT_PEER_CONNECTIONS,
-                            "daemon peer listener rejected a connection because the bounded concurrency cap was reached"
-                        );
-                        observability.emit_or_warn(
-                            "peer_connection_cap",
-                            "degraded",
-                            "daemon peer listener rejected a connection because the bounded concurrency cap was reached",
-                        );
-                        continue;
-                    }
-                    let dispatcher = Arc::clone(&dispatcher);
-                    let terminate = Arc::clone(&terminate);
-                    let observability = observability.clone();
-                    let registry = Arc::clone(&registry);
-                    scope.spawn(move || {
-                        let codec = JsonAtmProtocolCodec;
-                        let _connection = registry.register();
-                        let result = catch_unwind(AssertUnwindSafe(|| {
-                            handle_peer_connection(
-                                stream,
-                                dispatcher,
-                                &codec,
-                                terminate,
-                                registry,
-                            )
-                        }));
-                        match result {
-                            Ok(Ok(())) => {}
-                            Ok(Err(error)) => emit_peer_connection_failure_event(
-                                &observability,
-                                &error,
-                                None,
-                                "peer_connection",
-                            ),
-                            Err(_) => {
-                                let error = AtmError::daemon_unavailable(
-                                    "daemon peer listener worker panicked while handling a peer connection",
-                                )
-                                .with_recovery(
-                                    "Restart atm-daemon after reviewing the retained log for the panicking peer request.",
-                                );
-                                emit_peer_connection_failure_event(
-                                    &observability,
-                                    &error,
-                                    None,
-                                    "peer_connection",
-                                );
-                            }
-                        }
-                    });
-                }
-                Err(error) if error.kind() == io::ErrorKind::WouldBlock => {
-                    thread::sleep(PEER_LISTENER_ACCEPT_POLL_INTERVAL);
-                }
-                Err(error) => {
-                    tracing::error!(
-                        subsystem = "peer_transport",
-                        action = "accept_connection",
-                        outcome = "retrying",
-                        kind = ?error.kind(),
-                        %error,
-                        "daemon peer listener accept failed; keeping the listener alive and retrying"
-                    );
-                    observability.emit_or_warn(
-                        "peer_listener_accept",
-                        "degraded",
-                        "daemon peer listener accept failed; keeping the listener alive and retrying",
-                    );
-                    thread::sleep(PEER_ACCEPT_ERROR_RETRY_BACKOFF);
-                }
-            }
-        }
-        registry.interrupt_all();
-        registry.join_tracked_dispatches(PEER_REQUEST_DEADLINE)
-    })
-}
-
-fn handle_peer_connection(
-    mut stream: TcpStream,
-    dispatcher: Arc<dyn RequestDispatcher + Send + Sync>,
-    codec: &JsonAtmProtocolCodec,
-    terminate: Arc<AtomicBool>,
-    registry: Arc<ActiveConnectionRegistry>,
-) -> Result<(), AtmError> {
-    let deadline = Instant::now() + PEER_REQUEST_DEADLINE;
-    let Some(header_bytes) = read_peer_frame_header_until(
-        &mut stream,
-        deadline,
-        terminate.as_ref(),
-        "failed to read remote peer request frame",
-    )?
-    else {
-        return Ok(());
-    };
-    let header = atm_core::protocol::decode_frame_header(
-        header_bytes,
-        "remote peer request frame exceeded the maximum supported size",
-    )?;
-    let frame = read_peer_frame_payload_until(
-        &mut stream,
-        header,
-        deadline,
-        terminate.as_ref(),
-        "failed to read remote peer request frame",
-    )?;
-    let request_id = frame.request_id;
-    let (_, request) = codec.request_from_frame(frame)?;
-    let response = dispatch_peer_request(request, dispatcher, registry)?;
-    let response_frame = codec.response_to_frame(request_id, response)?;
-    write_peer_frame_until(
-        &mut stream,
-        &response_frame,
-        deadline,
-        terminate.as_ref(),
-        "failed to write remote peer response frame",
-    )?;
-    stream.flush().map_err(|source| {
-        AtmError::daemon_unavailable("failed to flush remote peer response frame")
-            .with_recovery("Retry after the remote daemon reconnects to the peer listener.")
-            .with_source(source)
-    })?;
-    Ok(())
-}
-
-fn dispatch_peer_request(
-    request: RequestEnvelope,
-    dispatcher: Arc<dyn RequestDispatcher + Send + Sync>,
-    registry: Arc<ActiveConnectionRegistry>,
-) -> Result<ResponseEnvelope, AtmError> {
-    let (result_tx, result_rx): (SyncSender<Result<ResponseEnvelope, AtmError>>, Receiver<_>) =
-        mpsc::sync_channel(1);
-    let (completion_tx, completion_rx) = mpsc::sync_channel(1);
-    let dispatch_registry = Arc::clone(&registry);
-    let join_handle = thread::Builder::new()
-        .name("peer-transport-dispatch".to_string())
-        .spawn(move || {
-            let _dispatch = dispatch_registry.register_dispatch_work();
-            let result = catch_unwind(AssertUnwindSafe(|| dispatcher.dispatch(request)));
-            let _ = match result {
-                Ok(result) => result_tx.send(result),
-                Err(_) => result_tx.send(Err(AtmError::daemon_unavailable(
-                    "daemon peer dispatch worker panicked unexpectedly",
-                )
-                .with_recovery(
-                    "Restart atm-daemon after confirming the destination daemon can complete peer request dispatches safely.",
-                ))),
-            };
-            let _ = completion_tx.send(());
-        })
-        .map_err(|source| {
-            AtmError::daemon_unavailable("failed to spawn daemon peer dispatch worker")
-                .with_recovery(
-                    "Restart atm-daemon after confirming the host can create cross-host peer dispatch workers.",
-                )
-                .with_source(source)
-        })?;
-    registry.push_dispatch_handle(
-        TrackedDispatchHandle {
-            completion_rx,
-            join_handle,
-        },
-        MAX_TRACKED_PEER_DISPATCH_HANDLES,
-    )?;
-    match result_rx.recv_timeout(PEER_REQUEST_DEADLINE) {
-        Ok(result) => result,
-        Err(mpsc::RecvTimeoutError::Timeout) => Err(AtmError::daemon_unavailable(
-            "remote peer request dispatch exceeded the bounded daemon deadline",
-        )
-        .with_recovery(
-            "Retry the cross-host peer operation after confirming the destination daemon is healthy and not stalled on request dispatch.",
-        )),
-        Err(mpsc::RecvTimeoutError::Disconnected) => Err(
-            AtmError::daemon_unavailable("daemon peer dispatch worker disconnected unexpectedly")
-                .with_recovery(
-                    "Restart atm-daemon after confirming the destination daemon can complete peer request dispatches.",
-                ),
-        ),
-    }
-}
-
-fn emit_peer_connection_failure_event(
-    observability: &SubsystemObservability,
-    error: &AtmError,
-    request_id: Option<atm_core::protocol::RequestId>,
-    transport_context: &'static str,
-) {
-    let classification = if error.is_validation() {
-        atm_core::observability::ConnectionFailureClassification::MalformedRequest
-    } else if error.is_daemon_unavailable() || error.is_timeout() {
-        atm_core::observability::ConnectionFailureClassification::TransportFailure
-    } else {
-        atm_core::observability::ConnectionFailureClassification::RequestFailure
-    };
-    observability.emit_event_or_warn(
-        observability
-            .event(
-                "peer_connection_worker",
-                classification.as_str(),
-                error.message.clone(),
-            )
-            .with_connection_failure(atm_core::observability::DaemonConnectionFailureFields {
-                code: error.code,
-                request_id,
-                classification,
-            })
-            .with_transport_context(transport_context),
-    );
-}
-
-fn read_peer_frame_header_until(
-    stream: &mut TcpStream,
-    deadline: Instant,
-    terminate: &AtomicBool,
-    read_error: &'static str,
-) -> Result<Option<[u8; atm_core::protocol::ATM_FRAME_HEADER_BYTES]>, AtmError> {
-    let mut header = [0u8; atm_core::protocol::ATM_FRAME_HEADER_BYTES];
-    let read = read_with_deadline(
-        stream,
-        &mut header[..1],
-        deadline,
-        terminate,
-        read_error,
-        true,
-    )?;
-    if read == 0 {
-        return Ok(None);
-    }
-    read_exact_with_deadline(stream, &mut header[1..], deadline, terminate, read_error)?;
-    Ok(Some(header))
-}
-
-fn read_peer_frame_payload_until(
-    stream: &mut TcpStream,
-    header: atm_core::protocol::FrameHeader,
-    deadline: Instant,
-    terminate: &AtomicBool,
-    read_error: &'static str,
-) -> Result<atm_core::protocol::FramePayload, AtmError> {
-    let mut bytes = vec![0u8; header.payload_length];
-    read_exact_with_deadline(stream, &mut bytes, deadline, terminate, read_error)?;
-    Ok(atm_core::protocol::FramePayload {
-        request_id: header.request_id,
-        message_kind: header.message_kind,
-        flags: header.flags,
-        bytes,
-    })
-}
-
-fn write_peer_frame_until(
-    stream: &mut TcpStream,
-    frame: &atm_core::protocol::FramePayload,
-    deadline: Instant,
-    terminate: &AtomicBool,
-    write_error: &'static str,
-) -> Result<(), AtmError> {
-    if frame.flags != atm_core::protocol::ATM_FRAME_FLAGS_V1 {
-        return Err(AtmError::validation(format!(
-            "unsupported ATM daemon frame flags 0x{:04x} for version {}",
-            frame.flags,
-            atm_core::protocol::ATM_FRAME_VERSION_V1
-        ))
-        .with_recovery(
-            "Retry with a supported ATM daemon client/server build that uses protocol version 1 flags.",
-        ));
-    }
-    if frame.bytes.len() > atm_core::protocol::MAX_DAEMON_FRAME_BYTES {
-        return Err(AtmError::daemon_unavailable(
-            "daemon frame exceeded the maximum supported size",
-        )
-        .with_recovery(
-            "Reduce the daemon request/response payload size before retrying the ATM command.",
-        ));
-    }
-    let mut header = [0u8; atm_core::protocol::ATM_FRAME_HEADER_BYTES];
-    header[0..4].copy_from_slice(&atm_core::protocol::ATM_FRAME_MAGIC.to_be_bytes());
-    header[4..6].copy_from_slice(&atm_core::protocol::ATM_FRAME_VERSION_V1.to_be_bytes());
-    header[6..8].copy_from_slice(&frame.message_kind.code().to_be_bytes());
-    header[8..10].copy_from_slice(&frame.flags.to_be_bytes());
-    header[10..18].copy_from_slice(&frame.request_id.into_inner().to_be_bytes());
-    header[18..22].copy_from_slice(&(frame.bytes.len() as u32).to_be_bytes());
-    write_all_with_deadline(stream, &header, deadline, terminate, write_error)?;
-    write_all_with_deadline(stream, &frame.bytes, deadline, terminate, write_error)
-}
-
-fn read_exact_with_deadline(
-    stream: &mut TcpStream,
-    mut buffer: &mut [u8],
-    deadline: Instant,
-    terminate: &AtomicBool,
-    read_error: &'static str,
-) -> Result<(), AtmError> {
-    while !buffer.is_empty() {
-        let read = read_with_deadline(stream, buffer, deadline, terminate, read_error, false)?;
-        if read == 0 {
-            return Err(AtmError::daemon_unavailable(read_error).with_recovery(
-                "Retry after the remote daemon reconnects and completes a bounded peer request/response exchange.",
-            ));
-        }
-        let (_, rest) = buffer.split_at_mut(read);
-        buffer = rest;
-    }
-    Ok(())
-}
-
-fn write_all_with_deadline(
-    stream: &mut TcpStream,
-    mut buffer: &[u8],
-    deadline: Instant,
-    terminate: &AtomicBool,
-    write_error: &'static str,
-) -> Result<(), AtmError> {
-    while !buffer.is_empty() {
-        apply_peer_io_slice_deadline(stream, deadline, terminate)?;
-        match stream.write(buffer) {
-            Ok(0) => {
-                return Err(AtmError::daemon_unavailable(write_error).with_recovery(
-                    "Retry after the remote daemon reconnects and completes a bounded peer request/response exchange.",
-                ));
-            }
-            Ok(written) => buffer = &buffer[written..],
-            Err(error)
-                if matches!(
-                    error.kind(),
-                    io::ErrorKind::TimedOut
-                        | io::ErrorKind::WouldBlock
-                        | io::ErrorKind::Interrupted
-                ) =>
-            {
-                continue;
-            }
-            Err(source) => {
-                return Err(AtmError::daemon_unavailable(write_error)
-                    .with_recovery("Retry after the remote daemon reconnects to the peer listener.")
-                    .with_source(source));
-            }
-        }
-    }
-    Ok(())
-}
-
-fn read_with_deadline(
-    stream: &mut TcpStream,
-    buffer: &mut [u8],
-    deadline: Instant,
-    terminate: &AtomicBool,
-    read_error: &'static str,
-    allow_eof: bool,
-) -> Result<usize, AtmError> {
-    loop {
-        apply_peer_io_slice_deadline(stream, deadline, terminate)?;
-        match stream.read(buffer) {
-            Ok(0) if allow_eof => return Ok(0),
-            Ok(0) => {
-                return Err(AtmError::daemon_unavailable(read_error).with_recovery(
-                    "Retry after the remote daemon reconnects and completes a bounded peer request/response exchange.",
-                ));
-            }
-            Ok(read) => return Ok(read),
-            Err(error)
-                if matches!(
-                    error.kind(),
-                    io::ErrorKind::TimedOut
-                        | io::ErrorKind::WouldBlock
-                        | io::ErrorKind::Interrupted
-                ) =>
-            {
-                continue;
-            }
-            Err(source) => {
-                return Err(AtmError::daemon_unavailable(read_error)
-                    .with_recovery("Retry after the remote daemon reconnects to the peer listener.")
-                    .with_source(source));
-            }
-        }
-    }
-}
-
-fn apply_peer_io_slice_deadline(
-    stream: &TcpStream,
-    deadline: Instant,
-    terminate: &AtomicBool,
-) -> Result<(), AtmError> {
-    if terminate.load(Ordering::SeqCst) {
-        return Err(AtmError::daemon_unavailable(
-            "daemon shutdown interrupted an in-flight peer connection",
-        )
-        .with_recovery(
-            "Retry the cross-host operation after atm-daemon restarts and resumes pending remote replay work.",
-        ));
-    }
-    let remaining = deadline.checked_duration_since(Instant::now()).ok_or_else(|| {
-        AtmError::daemon_unavailable("remote peer connection exceeded the bounded lifetime deadline")
-            .with_recovery(
-                "Retry the cross-host peer operation after confirming the destination daemon is healthy and not stalled on request handling.",
-            )
-    })?;
-    let slice = remaining.min(PEER_CONNECTION_IO_SLICE);
-    stream.set_read_timeout(Some(slice)).map_err(|source| {
-        AtmError::daemon_unavailable("failed to apply remote peer read deadline")
-            .with_recovery(
-                "Restart atm-daemon after confirming the host permits bounded TCP read deadlines for peer transport.",
-            )
-            .with_source(source)
-    })?;
-    stream.set_write_timeout(Some(slice)).map_err(|source| {
-        AtmError::daemon_unavailable("failed to apply remote peer write deadline")
-            .with_recovery(
-                "Restart atm-daemon after confirming the host permits bounded TCP write deadlines for peer transport.",
-            )
-            .with_source(source)
-    })?;
-    Ok(())
-}
-
-fn classify_io_error(error: &io::Error) -> AttemptFailureKind {
-    match error.kind() {
-        io::ErrorKind::TimedOut
-        | io::ErrorKind::Interrupted
-        | io::ErrorKind::ConnectionRefused
-        | io::ErrorKind::ConnectionReset
-        | io::ErrorKind::ConnectionAborted
-        | io::ErrorKind::BrokenPipe
-        | io::ErrorKind::AddrNotAvailable
-        | io::ErrorKind::NotConnected
-        | io::ErrorKind::HostUnreachable
-        | io::ErrorKind::NetworkUnreachable => AttemptFailureKind::Retryable,
-        _ => AttemptFailureKind::NonRetryable,
-    }
-}
-
-fn replay_metadata_for_request(
-    request: &RequestEnvelope,
-) -> Option<(TeamName, AgentName, MessageKey)> {
-    match request {
-        RequestEnvelope::Heartbeat(heartbeat) => Some((
-            heartbeat.team.clone(),
-            heartbeat.member.clone(),
-            heartbeat_message_key(heartbeat),
-        )),
-        _ => None,
-    }
-}
-
-fn heartbeat_message_key(request: &TeamMemberHeartbeatRequest) -> MessageKey {
-    // Team/member names are already validated ATM identifiers, pid is numeric, and RFC3339
-    // timestamps are never blank, so this synthetic replay key cannot violate MessageKey's
-    // non-empty invariant.
-    MessageKey::new(format!(
-        "remote-heartbeat:{}:{}:{}:{}",
-        request.team.as_str(),
-        request.member.as_str(),
-        request.pid,
-        request.observed_at.into_inner().to_rfc3339(),
-    ))
-    .expect("validated heartbeat replay keys are never blank")
-}
-
-fn jittered_backoff(base: Duration, seed: u64) -> Duration {
-    let base_nanos = base.as_nanos();
-    if base_nanos == 0 {
-        return Duration::ZERO;
-    }
-    let offset_basis_points = (seed % 4001) as i128 - 2000;
-    let scaled = (base_nanos as i128 * (10_000 + offset_basis_points))
-        .div_euclid(10_000)
-        .max(1);
-    Duration::from_nanos(u64::try_from(scaled).unwrap_or(u64::MAX))
-}
-
-fn jitter_seed(endpoint: SocketAddr, attempt: u32) -> u64 {
-    let now = match std::time::SystemTime::now().duration_since(std::time::UNIX_EPOCH) {
-        Ok(duration) => duration.as_nanos() as u64,
-        Err(error) => {
-            tracing::warn!(
-                subsystem = "peer_transport",
-                action = "jitter_seed",
-                outcome = "default",
-                %error,
-                "system clock is before the unix epoch; using deterministic jitter fallback"
-            );
-            0
-        }
-    };
-    now ^ u64::from(endpoint.port()) ^ (u64::from(attempt) << 32)
-}
-
-fn parse_peer_endpoint(raw: &str) -> Option<SocketAddr> {
-    match raw.parse::<SocketAddr>() {
-        Ok(endpoint) => Some(endpoint),
-        Err(error) => {
-            tracing::warn!(
-                subsystem = "peer_transport",
-                action = "parse_endpoint",
-                outcome = "skipped",
-                %raw,
-                %error,
-                "parse_peer_endpoint: invalid address format"
-            );
-            None
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        AttemptFailureKind, PeerClientTransport, PeerTransportConfig, PeerTransportRuntime,
-        classify_io_error, jittered_backoff, remote_peer_endpoint_not_configured_error,
-        remote_replay_persistence_failed_error, remote_replay_store_not_configured_error,
-    };
-    use crate::lifecycle_control::LifecycleControlSourceAdapter;
-    use crate::test_support::DoctorOnlyDispatcher;
-    use crate::test_support::LifecycleFlagResetGuard;
-    use crate::{DaemonSubsystem, SubsystemObservability};
-    use atm_core::boundary::{AtmProtocol, ClientTransport, MessageKey};
-    use atm_core::doctor::DoctorQuery;
-    use atm_core::error::AtmErrorCode;
-    use atm_core::protocol::{
-        HeartbeatActivity, ProtocolErrorEnvelope, RequestEnvelope, ResponseEnvelope,
-        RuntimeMemberState, TeamMemberHeartbeatRequest, TeamMemberHeartbeatResponse,
-    };
-    use atm_core::types::{AgentName, IsoTimestamp, TeamName};
-    use std::io::{self, Read, Write};
-    use std::net::{Ipv4Addr, SocketAddr, TcpListener, TcpStream};
-    use std::sync::Arc;
-    use std::sync::mpsc;
-    use std::thread;
-    use std::time::Duration;
-    use tempfile::TempDir;
-
-    fn read_request_frame(
-        stream: &mut TcpStream,
-    ) -> (
-        atm_core::protocol::RequestId,
-        RequestEnvelope,
-        super::JsonAtmProtocolCodec,
-    ) {
-        let codec = super::JsonAtmProtocolCodec;
-        let frame = atm_core::protocol::read_frame(
-            stream,
-            "read request",
-            "request frame exceeded frame limit",
-        )
-        .expect("read frame")
-        .expect("request frame");
-        let (request_id, request) = codec.request_from_frame(frame).expect("decode request");
-        (request_id, request, codec)
-    }
-
-    fn write_response_frame(
-        stream: &mut TcpStream,
-        codec: &super::JsonAtmProtocolCodec,
-        request_id: atm_core::protocol::RequestId,
-        response: ResponseEnvelope,
-    ) {
-        let frame = codec
-            .response_to_frame(request_id, response)
-            .expect("response frame");
-        atm_core::protocol::write_frame(stream, &frame, "write response").expect("write response");
-        stream.flush().expect("flush response");
-    }
-
-    fn install_shared_lifecycle_reset_guard() -> LifecycleFlagResetGuard {
-        let lifecycle = LifecycleControlSourceAdapter::install().expect("install lifecycle");
-        LifecycleFlagResetGuard::install(lifecycle)
-    }
-
-    #[test]
-    fn peer_listener_round_trips_one_doctor_request() {
-        let _guard = install_shared_lifecycle_reset_guard();
-        let listener_transport = PeerTransportRuntime::new_server_for_test(
-            SocketAddr::from((Ipv4Addr::LOCALHOST, 0)),
-            SubsystemObservability::disabled(DaemonSubsystem::PeerTransport),
-        );
-        listener_transport
-            .start(Arc::new(DoctorOnlyDispatcher))
-            .expect("start peer listener");
-        let endpoint = listener_transport
-            .bound_addr_for_test()
-            .expect("bound peer listener addr");
-
-        let tempdir = TempDir::new().expect("tempdir");
-        let client_transport = PeerTransportRuntime::new_for_test(
-            endpoint,
-            PeerTransportConfig::default(),
-            tempdir.path().join("replay.db"),
-        );
-        let response = client_transport
-            .client_transport()
-            .send(RequestEnvelope::Doctor(atm_core::doctor::DoctorQuery {
-                home_dir: tempdir.path().to_path_buf(),
-                current_dir: tempdir.path().to_path_buf(),
-                team_override: None,
-            }))
-            .expect("peer listener doctor response");
-        match response {
-            ResponseEnvelope::Doctor(report) => {
-                assert_eq!(
-                    report.summary.status,
-                    atm_core::doctor::DoctorStatus::Healthy
-                );
-            }
-            other => panic!("unexpected response from peer listener: {other:?}"),
-        }
-
-        listener_transport
-            .shutdown()
-            .expect("shutdown peer listener");
-    }
-
-    #[test]
-    fn jittered_backoff_stays_within_twenty_percent_window() {
-        let base = Duration::from_millis(250);
-        let low = jittered_backoff(base, 0);
-        let high = jittered_backoff(base, 4_000);
-        assert_eq!(low, Duration::from_millis(200));
-        assert_eq!(high, Duration::from_millis(300));
-    }
-
-    #[test]
-    fn classify_io_error_covers_retryable_and_non_retryable_variants() {
-        for kind in [
-            io::ErrorKind::TimedOut,
-            io::ErrorKind::Interrupted,
-            io::ErrorKind::ConnectionRefused,
-            io::ErrorKind::ConnectionReset,
-            io::ErrorKind::ConnectionAborted,
-            io::ErrorKind::BrokenPipe,
-            io::ErrorKind::HostUnreachable,
-        ] {
-            assert_eq!(
-                classify_io_error(&io::Error::new(kind, "retryable")),
-                AttemptFailureKind::Retryable
-            );
-        }
-        assert_eq!(
-            classify_io_error(&io::Error::other("non-retryable")),
-            AttemptFailureKind::NonRetryable
-        );
-    }
-
-    #[test]
-    fn persist_replay_request_requires_configured_replay_store_with_recovery() {
-        let team: TeamName = "test-team".parse().expect("team");
-        let agent: AgentName = "test-agent".parse().expect("agent");
-        let client = PeerClientTransport {
-            endpoint: Some(SocketAddr::from((Ipv4Addr::LOCALHOST, 7001))),
-            config: PeerTransportConfig::default(),
-            replay_store: None,
-            codec: super::JsonAtmProtocolCodec,
-            observability: SubsystemObservability::disabled(DaemonSubsystem::PeerTransport),
-        };
-        let error = client
-            .persist_replay_request(
-                team.clone(),
-                agent.clone(),
-                MessageKey::new("atm:test-remote-replay-store-missing").expect("message key"),
-                RequestEnvelope::Heartbeat(TeamMemberHeartbeatRequest {
-                    team,
-                    member: agent,
-                    pid: 42,
-                    observed_at: IsoTimestamp::now(),
-                    activity: HeartbeatActivity::Idle,
-                }),
-            )
-            .expect_err("missing replay store should fail closed");
-        assert_eq!(error.code, AtmErrorCode::DaemonUnavailable);
-        assert!(
-            error
-                .primary_recovery()
-                .expect("recovery guidance")
-                .contains("atm-daemon binary is installed")
-        );
-    }
-
-    #[test]
-    fn persist_replay_request_missing_endpoint_matches_send_surface_contract() {
-        let tempdir = TempDir::new().expect("tempdir");
-        let team: TeamName = "test-team".parse().expect("team");
-        let agent: AgentName = "test-agent".parse().expect("agent");
-        let replay_store =
-            atm_runtime::sqlite_remote_replay_store_for_test(tempdir.path().join("mail.db"))
-                .expect("replay store");
-        let client = PeerClientTransport {
-            endpoint: None,
-            config: PeerTransportConfig::default(),
-            replay_store: Some(replay_store),
-            codec: super::JsonAtmProtocolCodec,
-            observability: SubsystemObservability::disabled(DaemonSubsystem::PeerTransport),
-        };
-        let persist_error = client
-            .persist_replay_request(
-                team.clone(),
-                agent.clone(),
-                MessageKey::new("atm:test-remote-peer-endpoint-missing").expect("message key"),
-                RequestEnvelope::Heartbeat(TeamMemberHeartbeatRequest {
-                    team,
-                    member: agent,
-                    pid: 43,
-                    observed_at: IsoTimestamp::now(),
-                    activity: HeartbeatActivity::Idle,
-                }),
-            )
-            .expect_err("missing endpoint should fail closed");
-        let send_error = remote_peer_endpoint_not_configured_error();
-        assert_eq!(persist_error.code, send_error.code);
-        assert_eq!(persist_error.recovery, send_error.recovery);
-    }
-
-    #[test]
-    fn persist_replay_request_invalid_retry_budget_reports_actionable_recovery() {
-        let tempdir = TempDir::new().expect("tempdir");
-        let team: TeamName = "test-team".parse().expect("team");
-        let agent: AgentName = "test-agent".parse().expect("agent");
-        let replay_store =
-            atm_runtime::sqlite_remote_replay_store_for_test(tempdir.path().join("mail.db"))
-                .expect("replay store");
-        let client = PeerClientTransport {
-            endpoint: Some(SocketAddr::from((Ipv4Addr::LOCALHOST, 7002))),
-            config: PeerTransportConfig {
-                remote_retry_budget: Duration::MAX,
-                peer_listen_addr: None,
-            },
-            replay_store: Some(replay_store),
-            codec: super::JsonAtmProtocolCodec,
-            observability: SubsystemObservability::disabled(DaemonSubsystem::PeerTransport),
-        };
-        let error = client
-            .persist_replay_request(
-                team.clone(),
-                agent.clone(),
-                MessageKey::new("atm:test-remote-retry-budget-invalid").expect("message key"),
-                RequestEnvelope::Heartbeat(TeamMemberHeartbeatRequest {
-                    team,
-                    member: agent,
-                    pid: 44,
-                    observed_at: IsoTimestamp::now(),
-                    activity: HeartbeatActivity::Idle,
-                }),
-            )
-            .expect_err("invalid retry budget should fail closed");
-        assert_eq!(error.code, AtmErrorCode::DaemonUnavailable);
-        assert!(
-            error
-                .primary_recovery()
-                .expect("recovery guidance")
-                .contains("atm-daemon binary is installed")
-        );
-    }
-
-    #[test]
-    fn protocol_envelope_preserves_replay_persistence_failure_contract() {
-        let error =
-            remote_replay_persistence_failed_error(remote_replay_store_not_configured_error());
-        let envelope = ProtocolErrorEnvelope::from_error(&error);
-        let round_trip = envelope.into_atm_error();
-        assert_eq!(round_trip.code, AtmErrorCode::RemoteDeliveryOutcomeUnknown);
-        assert_eq!(round_trip.message, error.message);
-        assert_eq!(round_trip.recovery, error.recovery);
-    }
-
-    #[test]
-    fn wildcard_bindings_survive_connection_churn_and_explicit_binds_require_reload() {
-        let listener = TcpListener::bind((Ipv4Addr::UNSPECIFIED, 0)).expect("wildcard listener");
-        let endpoint = SocketAddr::from((
-            Ipv4Addr::LOCALHOST,
-            listener.local_addr().expect("addr").port(),
-        ));
-
-        let server = thread::spawn(move || {
-            for _ in 0..2 {
-                let (mut stream, _) = listener.accept().expect("accept");
-                let mut buffer = [0u8; 1];
-                stream.read_exact(&mut buffer).expect("read");
-            }
-        });
-
-        for _ in 0..2 {
-            let mut stream = TcpStream::connect(endpoint).expect("connect");
-            stream.write_all(&[1]).expect("write");
-        }
-        server.join().expect("server join");
-
-        let bind_error = TcpListener::bind((Ipv4Addr::new(198, 51, 100, 10), 0))
-            .expect_err("explicit bind failure");
-        assert_eq!(bind_error.kind(), io::ErrorKind::AddrNotAvailable);
-    }
-
-    #[test]
-    #[serial_test::serial(env)]
-    fn peer_transport_round_trips_one_heartbeat_request() {
-        let _reset = install_shared_lifecycle_reset_guard();
-        let tempdir = TempDir::new().expect("tempdir");
-        let listener = TcpListener::bind((Ipv4Addr::LOCALHOST, 0)).expect("listener");
-        let endpoint = listener.local_addr().expect("addr");
-        let transport = PeerTransportRuntime::new_for_test(
-            endpoint,
-            PeerTransportConfig::default(),
-            tempdir.path().join("mail.db"),
-        );
-        let team: TeamName = "test-team".parse().expect("team");
-        let member: AgentName = "test-member".parse().expect("member");
-        let request = RequestEnvelope::Heartbeat(TeamMemberHeartbeatRequest {
-            team: team.clone(),
-            member: member.clone(),
-            pid: 42,
-            observed_at: IsoTimestamp::now(),
-            activity: HeartbeatActivity::Idle,
-        });
-        let expected = ResponseEnvelope::Heartbeat(TeamMemberHeartbeatResponse {
-            team,
-            member,
-            pid: 42,
-            pid_changed: false,
-            state: RuntimeMemberState::Idle,
-            last_active_at: Some(IsoTimestamp::now()),
-        });
-
-        thread::spawn(move || {
-            let (mut stream, _) = listener.accept().expect("accept");
-            let (request_id, _request, codec) = read_request_frame(&mut stream);
-            write_response_frame(&mut stream, &codec, request_id, expected);
-        });
-
-        let response = transport
-            .client_transport()
-            .send(request)
-            .expect("response");
-        match response {
-            ResponseEnvelope::Heartbeat(response) => {
-                assert_eq!(response.state, RuntimeMemberState::Idle)
-            }
-            other => panic!("unexpected response: {other:?}"),
-        }
-    }
-
-    #[test]
-    #[serial_test::serial(env)]
-    fn peer_transport_aborts_before_connect_when_terminate_is_requested() {
-        let _reset = install_shared_lifecycle_reset_guard();
-        let lifecycle = LifecycleControlSourceAdapter::install().expect("install lifecycle");
-        lifecycle.set_terminate_for_test(true);
-
-        let tempdir = TempDir::new().expect("tempdir");
-        let endpoint = {
-            let listener = TcpListener::bind((Ipv4Addr::LOCALHOST, 0)).expect("listener");
-            let endpoint = listener.local_addr().expect("addr");
-            drop(listener);
-            endpoint
-        };
-        let transport = PeerTransportRuntime::new_for_test(
-            endpoint,
-            PeerTransportConfig {
-                remote_retry_budget: Duration::from_secs(1),
-                peer_listen_addr: None,
-            },
-            tempdir.path().join("replay.db"),
-        );
-
-        let error = transport
-            .client_transport()
-            .send(RequestEnvelope::Heartbeat(TeamMemberHeartbeatRequest {
-                team: atm_core::test_support::TEST_TEAM.parse().expect("team"),
-                member: atm_core::test_support::TEST_SENDER.parse().expect("member"),
-                pid: std::process::id(),
-                observed_at: IsoTimestamp::now(),
-                activity: HeartbeatActivity::ActiveToolUse,
-            }))
-            .expect_err("terminate should short-circuit before connect");
-        assert_eq!(error.code, AtmErrorCode::DaemonUnavailable);
-    }
-
-    #[test]
-    #[serial_test::serial(env)]
-    fn peer_transport_uses_port_zero_listener_handoff_without_rebind_race() {
-        let _reset = install_shared_lifecycle_reset_guard();
-        let tempdir = TempDir::new().expect("tempdir");
-        let (endpoint_tx, endpoint_rx) = mpsc::channel();
-        let team: TeamName = "test-team".parse().expect("team");
-        let member: AgentName = "test-member".parse().expect("member");
-        let server_team = team.clone();
-        let server_member = member.clone();
-
-        thread::spawn(move || {
-            let listener = TcpListener::bind((Ipv4Addr::LOCALHOST, 0)).expect("listener");
-            endpoint_tx
-                .send(listener.local_addr().expect("addr"))
-                .expect("endpoint");
-            let (mut stream, _) = listener.accept().expect("accept");
-            let (request_id, _request, codec) = read_request_frame(&mut stream);
-            let response = ResponseEnvelope::Heartbeat(TeamMemberHeartbeatResponse {
-                team: server_team,
-                member: server_member,
-                pid: 7,
-                pid_changed: false,
-                state: RuntimeMemberState::Active,
-                last_active_at: Some(IsoTimestamp::now()),
-            });
-            write_response_frame(&mut stream, &codec, request_id, response);
-        });
-
-        let endpoint = endpoint_rx.recv().expect("endpoint");
-
-        let transport = PeerTransportRuntime::new_for_test(
-            endpoint,
-            PeerTransportConfig {
-                remote_retry_budget: Duration::from_millis(600),
-                peer_listen_addr: None,
-            },
-            tempdir.path().join("mail.db"),
-        );
-        let request = RequestEnvelope::Heartbeat(TeamMemberHeartbeatRequest {
-            team: team.clone(),
-            member: member.clone(),
-            pid: 7,
-            observed_at: IsoTimestamp::now(),
-            activity: HeartbeatActivity::ActiveToolUse,
-        });
-        let (send_started_tx, send_started_rx) = mpsc::channel();
-        let (response_tx, response_rx) = mpsc::channel();
-
-        let transport_for_thread = transport.clone();
-        thread::spawn(move || {
-            send_started_tx.send(()).expect("send started");
-            response_tx
-                .send(transport_for_thread.client.send(request))
-                .expect("response sent");
-        });
-
-        send_started_rx.recv().expect("send started");
-        let response = response_rx
-            .recv_timeout(Duration::from_secs(5))
-            .expect("response wait")
-            .expect("response delivered");
-        assert!(matches!(response, ResponseEnvelope::Heartbeat(_)));
-    }
-
-    #[test]
-    #[serial_test::serial(env)]
-    fn peer_transport_reports_outcome_unknown_after_send_without_response() {
-        let _reset = install_shared_lifecycle_reset_guard();
-        let tempdir = TempDir::new().expect("tempdir");
-        let listener = TcpListener::bind((Ipv4Addr::LOCALHOST, 0)).expect("listener");
-        let endpoint = listener.local_addr().expect("addr");
-        let transport = PeerTransportRuntime::new_for_test(
-            endpoint,
-            PeerTransportConfig::default(),
-            tempdir.path().join("mail.db"),
-        );
-        let team: TeamName = "test-team".parse().expect("team");
-        let member: AgentName = "test-member".parse().expect("member");
-        let request = RequestEnvelope::Heartbeat(TeamMemberHeartbeatRequest {
-            team,
-            member,
-            pid: 11,
-            observed_at: IsoTimestamp::now(),
-            activity: HeartbeatActivity::Idle,
-        });
-
-        thread::spawn(move || {
-            let (mut stream, _) = listener.accept().expect("accept");
-            let _ = read_request_frame(&mut stream);
-        });
-
-        let error = transport
-            .client_transport()
-            .send(request)
-            .expect_err("outcome unknown");
-        assert_eq!(error.code, AtmErrorCode::RemoteDeliveryOutcomeUnknown);
-    }
-
-    #[test]
-    #[serial_test::serial(env)]
-    fn peer_transport_treats_remote_error_envelope_as_non_retryable() {
-        let _reset = install_shared_lifecycle_reset_guard();
-        let tempdir = TempDir::new().expect("tempdir");
-        let listener = TcpListener::bind((Ipv4Addr::LOCALHOST, 0)).expect("listener");
-        let endpoint = listener.local_addr().expect("addr");
-        let transport = PeerTransportRuntime::new_for_test(
-            endpoint,
-            PeerTransportConfig::default(),
-            tempdir.path().join("mail.db"),
-        );
-        let team: TeamName = "test-team".parse().expect("team");
-        let member: AgentName = "test-member".parse().expect("member");
-        let request = RequestEnvelope::Heartbeat(TeamMemberHeartbeatRequest {
-            team,
-            member,
-            pid: 12,
-            observed_at: IsoTimestamp::now(),
-            activity: HeartbeatActivity::Idle,
-        });
-
-        thread::spawn(move || {
-            let (mut stream, _) = listener.accept().expect("accept");
-            let (request_id, _request, codec) = read_request_frame(&mut stream);
-            let response = ResponseEnvelope::Error(ProtocolErrorEnvelope {
-                code: AtmErrorCode::DaemonUnavailable,
-                message: "remote rejected request".to_string(),
-                recovery: Vec::new(),
-            });
-            write_response_frame(&mut stream, &codec, request_id, response);
-        });
-
-        let error = transport
-            .client_transport()
-            .send(request)
-            .expect_err("remote reject");
-        assert_eq!(error.code, AtmErrorCode::DaemonUnavailable);
-    }
-
-    #[test]
-    #[serial_test::serial(env)]
-    fn replay_resume_replays_and_deletes_delivered_rows() {
-        let _reset = install_shared_lifecycle_reset_guard();
-        let tempdir = TempDir::new().expect("tempdir");
-        let listener = TcpListener::bind((Ipv4Addr::LOCALHOST, 0)).expect("listener");
-        let endpoint = listener.local_addr().expect("addr");
-        let db_path = tempdir.path().join("mail.db");
-        let transport = PeerTransportRuntime::new_for_test(
-            endpoint,
-            PeerTransportConfig::default(),
-            db_path.clone(),
-        );
-        let team: TeamName = "test-team".parse().expect("team");
-        let member: AgentName = "test-member".parse().expect("member");
-        let request = RequestEnvelope::Heartbeat(TeamMemberHeartbeatRequest {
-            team: team.clone(),
-            member: member.clone(),
-            pid: 21,
-            observed_at: IsoTimestamp::now(),
-            activity: HeartbeatActivity::Idle,
-        });
-        transport
-            .persist_replay_request(
-                team.clone(),
-                member.clone(),
-                MessageKey::new("atm:test-remote-replay").expect("message key"),
-                request,
-            )
-            .expect("persist");
-
-        thread::spawn(move || {
-            let (mut stream, _) = listener.accept().expect("accept");
-            let (request_id, _request, codec) = read_request_frame(&mut stream);
-            let response = ResponseEnvelope::Heartbeat(TeamMemberHeartbeatResponse {
-                team,
-                member,
-                pid: 21,
-                pid_changed: false,
-                state: RuntimeMemberState::Idle,
-                last_active_at: Some(IsoTimestamp::now()),
-            });
-            write_response_frame(&mut stream, &codec, request_id, response);
-        });
-
-        let summary = transport.resume_pending_replay().expect("resume");
-        assert_eq!(summary.delivered, 1);
-        assert_eq!(summary.retained, 0);
-        let pending = transport
-            .load_pending_replay_records()
-            .expect("load pending");
-        assert!(pending.is_empty());
-    }
-
-    #[test]
-    #[serial_test::serial(env)]
-    fn outcome_unknown_persists_replay_request_for_restart_resume() {
-        let _reset = install_shared_lifecycle_reset_guard();
-        let tempdir = TempDir::new().expect("tempdir");
-        let listener = TcpListener::bind((Ipv4Addr::LOCALHOST, 0)).expect("listener");
-        let endpoint = listener.local_addr().expect("addr");
-        let db_path = tempdir.path().join("mail.db");
-        let transport = PeerTransportRuntime::new_for_test(
-            endpoint,
-            PeerTransportConfig::default(),
-            db_path.clone(),
-        );
-        let team: TeamName = "test-team".parse().expect("team");
-        let member: AgentName = "test-member".parse().expect("member");
-        let request = RequestEnvelope::Heartbeat(TeamMemberHeartbeatRequest {
-            team: team.clone(),
-            member: member.clone(),
-            pid: 77,
-            observed_at: IsoTimestamp::now(),
-            activity: HeartbeatActivity::Idle,
-        });
-
-        thread::spawn(move || {
-            let (mut stream, _) = listener.accept().expect("accept");
-            let _ = read_request_frame(&mut stream);
-        });
-
-        let error = transport
-            .client_transport()
-            .send(request)
-            .expect_err("outcome unknown");
-        assert_eq!(error.code, AtmErrorCode::RemoteDeliveryOutcomeUnknown);
-
-        let pending = transport
-            .load_pending_replay_records()
-            .expect("load pending");
-        assert_eq!(pending.len(), 1);
-        assert_eq!(pending[0].team, team);
-        assert_eq!(pending[0].agent, member);
-    }
-
-    #[test]
-    #[serial_test::serial(env)]
-    fn unsupported_request_family_keeps_shared_outcome_unknown_recovery() {
-        let _reset = install_shared_lifecycle_reset_guard();
-        let tempdir = TempDir::new().expect("tempdir");
-        let listener = TcpListener::bind((Ipv4Addr::LOCALHOST, 0)).expect("listener");
-        let endpoint = listener.local_addr().expect("addr");
-        let db_path = tempdir.path().join("mail.db");
-        let transport = PeerTransportRuntime::new_for_test(
-            endpoint,
-            PeerTransportConfig::default(),
-            db_path.clone(),
-        );
-        let request = RequestEnvelope::Doctor(DoctorQuery {
-            home_dir: tempdir.path().to_path_buf(),
-            current_dir: tempdir.path().to_path_buf(),
-            team_override: None,
-        });
-
-        thread::spawn(move || {
-            let (mut stream, _) = listener.accept().expect("accept");
-            let _ = read_request_frame(&mut stream);
-        });
-
-        let error = transport
-            .client_transport()
-            .send(request)
-            .expect_err("outcome unknown");
-        assert_eq!(error.code, AtmErrorCode::RemoteDeliveryOutcomeUnknown);
-        assert!(
-            error
-                .primary_recovery()
-                .expect("recovery guidance")
-                .contains("let the daemon resume the pending handoff")
-        );
-        let pending = transport
-            .load_pending_replay_records()
-            .expect("load pending");
-        assert!(pending.is_empty());
-    }
-
-    #[test]
-    #[serial_test::serial(env)]
-    fn replay_resume_after_restart_delivers_once_and_clears_duplicate_delivery() {
-        let _reset = install_shared_lifecycle_reset_guard();
-        let tempdir = TempDir::new().expect("tempdir");
-        let listener = TcpListener::bind((Ipv4Addr::LOCALHOST, 0)).expect("listener");
-        let endpoint = listener.local_addr().expect("addr");
-        let db_path = tempdir.path().join("mail.db");
-        let first = PeerTransportRuntime::new_for_test(
-            endpoint,
-            PeerTransportConfig::default(),
-            db_path.clone(),
-        );
-        let second = PeerTransportRuntime::new_for_test(
-            endpoint,
-            PeerTransportConfig::default(),
-            db_path.clone(),
-        );
-        let team: TeamName = "test-team".parse().expect("team");
-        let member: AgentName = "test-member".parse().expect("member");
-        let request = RequestEnvelope::Heartbeat(TeamMemberHeartbeatRequest {
-            team: team.clone(),
-            member: member.clone(),
-            pid: 32,
-            observed_at: IsoTimestamp::now(),
-            activity: HeartbeatActivity::Idle,
-        });
-        first
-            .persist_replay_request(
-                team.clone(),
-                member.clone(),
-                MessageKey::new("atm:test-remote-replay-restart").expect("message key"),
-                request,
-            )
-            .expect("persist");
-
-        let (deliveries_tx, deliveries_rx) = mpsc::channel();
-        thread::spawn(move || {
-            let (mut stream, _) = listener.accept().expect("accept");
-            let (request_id, _request, codec) = read_request_frame(&mut stream);
-            deliveries_tx.send(()).expect("delivery sent");
-            let response = ResponseEnvelope::Heartbeat(TeamMemberHeartbeatResponse {
-                team,
-                member,
-                pid: 32,
-                pid_changed: false,
-                state: RuntimeMemberState::Idle,
-                last_active_at: Some(IsoTimestamp::now()),
-            });
-            write_response_frame(&mut stream, &codec, request_id, response);
-        });
-
-        let summary = second.resume_pending_replay().expect("resume");
-        assert_eq!(summary.delivered, 1);
-        deliveries_rx.recv().expect("delivery");
-
-        let pending = second.load_pending_replay_records().expect("load pending");
-        assert!(pending.is_empty());
-
-        let summary = second.resume_pending_replay().expect("second resume");
-        assert_eq!(summary.delivered, 0);
-        assert_eq!(summary.retained, 0);
-    }
-
-    #[test]
-    fn replay_store_upsert_deduplicates_same_message_key() {
-        let tempdir = TempDir::new().expect("tempdir");
-        let listener = TcpListener::bind((Ipv4Addr::LOCALHOST, 0)).expect("listener");
-        let endpoint = listener.local_addr().expect("addr");
-        drop(listener);
-        let transport = PeerTransportRuntime::new_for_test(
-            endpoint,
-            PeerTransportConfig::default(),
-            tempdir.path().join("mail.db"),
-        );
-        let team: TeamName = "test-team".parse().expect("team");
-        let member: AgentName = "test-member".parse().expect("member");
-        let message_key = MessageKey::new("atm:test-remote-replay-dedup").expect("message key");
-        let first_request = RequestEnvelope::Heartbeat(TeamMemberHeartbeatRequest {
-            team: team.clone(),
-            member: member.clone(),
-            pid: 11,
-            observed_at: IsoTimestamp::now(),
-            activity: HeartbeatActivity::Idle,
-        });
-        let second_request = RequestEnvelope::Heartbeat(TeamMemberHeartbeatRequest {
-            team: team.clone(),
-            member: member.clone(),
-            pid: 12,
-            observed_at: IsoTimestamp::now(),
-            activity: HeartbeatActivity::ActiveToolUse,
-        });
-
-        transport
-            .persist_replay_request(
-                team.clone(),
-                member.clone(),
-                message_key.clone(),
-                first_request,
-            )
-            .expect("persist first");
-        transport
-            .persist_replay_request(team, member, message_key, second_request)
-            .expect("persist second");
-
-        let pending = transport
-            .load_pending_replay_records()
-            .expect("load pending");
-        assert_eq!(pending.len(), 1);
-        let RequestEnvelope::Heartbeat(heartbeat) = &pending[0].request else {
-            panic!("expected heartbeat replay record");
-        };
-        assert_eq!(heartbeat.pid, 12);
-    }
-}
