+++ docs/plans/phase-AG/sprint-AG1.md
@@ -45,6 +45,11 @@ attempt the first live cross-host daemon-to-daemon channel.
   - `AG-VAL-011`
   - AG.1 viability may exercise `AG-VAL-003` or `AG-VAL-005`, but does not
     formally close them
+- peer-listener deliverable
+  - `cargo test -p atm-daemon peer_transport -- --nocapture`
+  - `cargo test -p atm-daemon -- --nocapture`
+  - evidence artifact: retained `docs/plans/phase-AG/reports/macos-report.md`
+    entry naming PR #551 / the validation commit used for the AG.1 rerun
 
 ## Ownership
 
@@ -57,3 +62,34 @@ attempt the first live cross-host daemon-to-daemon channel.
 - the runbook is concrete enough that both hosts can execute without guessing
 - the first live channel attempt has a defined pass/fail evidence contract
 - setup ambiguity is classified as a finding instead of being hand-waved away
+- the peer-listener fix is production-ready only if the inbound listener
+  starts, reload/rebind degradation is queryable in doctor/runtime status, and
+  the bounded peer transport tests above pass on the sprint branch
+
+## PeerServerTransport contract
+
+```rust
+pub(crate) struct PeerServerTransport {
+    listen_addr: Mutex<Option<SocketAddr>>,
+    observability: SubsystemObservability,
+    state: Mutex<Option<PeerServerHandle>>,
+    status_cache: RuntimeStatusCache,
+}
+
+impl PeerServerTransport {
+    fn start(&self, dispatcher: Arc<dyn RequestDispatcher + Send + Sync>)
+        -> Result<(), AtmError>;
+    fn shutdown(&self) -> Result<(), AtmError>;
+    fn reload(
+        &self,
+        listen_addr: Option<SocketAddr>,
+        dispatcher: Arc<dyn RequestDispatcher + Send + Sync>,
+    ) -> Result<(), AtmError>;
+}
+```
+
+Notes:
+
+- `listen_addr` comes from `daemon.peer_listen_addr`
+- `reload(...)` must preserve one bounded runtime view: if rebind fails, doctor
+  surfaces degraded listener state until a later successful rebind clears it
